SuctionPipe code.  Work code not in yet.

+++ mcp/src/minecraft/femtocraft/api/ISuctionPipe.java
@@ -13,5 +13,5 @@ public interface ISuctionPipe extends IFluidHandler {
 	 * 
 	 * @return Pressure of this pipe.  Liquids will flow from high pressure to low pressure pipes.
 	 */
-	public float getPressure();
+	public int getPressure();
 }
